% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Delta_transitioned}
\alias{Delta_transitioned}
\title{Rasterized and transitioned Delta shapefile}
\format{a TransitionLayer from package \link{gdistance}.}
\usage{
Delta_transitioned
}
\description{
Shapefile of delta waterways that has been rasterized and transitioned for Waterdist.
}
\keyword{datasets}
