% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGdist.R
\name{GGdist}
\alias{GGdist}
\title{In-water distances between a set of points and the Golden Gate}
\usage{
GGdist(
  Water_map,
  Points,
  EndPoint = NULL,
  Latitude_column,
  Longitude_column,
  PointID_column,
  Points_crs = 4326,
  Water_map_transitioned = NULL,
  Calculation_crs = 32610,
  Grid_size = 75
)
}
\arguments{
\item{Water_map}{Object of class sf representing a map of all waterways in your region of interest}

\item{Points}{A dataframe of points with latitude and longitude which you would like to calculate distance to a given reference point.}

\item{EndPoint}{A dataframe containing a single point with latitude and longitude. Defaults to a point 0.1 degrees East of the Golden Gate (37.819539, -122.378041). Latitude and longitude column names must be the same as the points data frame. This point must fall within the bounds of the \code{Water_map}.}

\item{Latitude_column}{The unquoted name of the column in the \code{Points} and \code{EndPoint} dataframe representing Latitude.}

\item{Longitude_column}{The unquoted name of the column in the \code{Points} and \code{EndPoint} dataframe representing Longitude.}

\item{PointID_column}{The unquoted name of the column in the \code{Points} dataframe with the unique identifier of each point.}

\item{Points_crs}{Coordinate reference system for your \code{Points} dataframe. Integer with the EPSG code or character with proj4string.}

\item{Water_map_transitioned}{A rasterized, transitioned, and geo-corrected version of the water map. This is optional to save time if you will be running this function frequently with the same base map.}

\item{Calculation_crs}{Coordinate reference system used for the calculation. If a latitude/longitude system are used, errors may be returned since some calculations assume a planar surface. Defaults to \code{Calculation_crs = 32610}.}

\item{Grid_size}{Grid size (in meters) used to rasterize the map. Defaults to 75.}
}
\value{
A tibble with 1 column for the PointID_column values and another with the distances from the EndPoint.
}
\description{
Calculate a distance matrix for a set of points based on in-water distances, using a raster-based approach.
}
\examples{
library(tibble)

Points <- tibble(Latitude = c(38.23333, 38.04813, 38.05920,
                                      37.94900, 38.23615, 38.47387),
                         Longitude = c(-121.4889, -121.9149, -121.8684,
                                       -121.5591, -121.6735, -121.5844),
                         ID = c("EMP NZP02", "FMWT 508", "FMWT 513",
                                "FMWT 915", "FMWT 723", "FMWT 796"))

\dontrun{
distance<GGdist(Water_map = spacetools::Delta, Points = Points, Latitude_column = Latitude,
                    Longitude_column = Longitude, PointID_column = ID)


# Including a pre-transitioned map to save time.
# See Maptransitioner for creating this pre-transitioned map.

distance<-GGdist(Water_map = spacetools::Delta, Points = Points, Latitude_column = Latitude,
                    Longitude_column = Longitude, PointID_column = ID,
                    Water_map_transitioned = Delta_transitioned)
                    }
}
\seealso{
\code{\link{Maptransitioner}} \code{\link{Pointmover}}
}
\keyword{distance}
\keyword{raster}
\keyword{spatial}
\keyword{water}
