% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maptransitioner.R
\name{Maptransitioner}
\alias{Maptransitioner}
\title{Rasterize and transition a shapefile}
\usage{
Maptransitioner(
  Water_map,
  Calculation_crs = "+proj=utm +zone=10 ellps=WGS84",
  Grid_size = 75,
  Process_map = TRUE,
  Plot = FALSE
)
}
\arguments{
\item{Water_map}{Object of class sf representing a map of all waterways in your region of interest}

\item{Calculation_crs}{Coordinate reference system used for the calculation. If a latitude/longitude system are used, errors may be returned since some calculations assume a planar surface. Defaults to \code{Calculation_crs = "+proj=utm +zone=10 ellps=WGS84"}.}

\item{Grid_size}{Grid size (in meters) used to rasterize the map. Defaults to 75.}

\item{Process_map}{Should the \code{Water_map} be processed by unioning all features and transforming to the \code{Calculation_crs}? This should almost always be set to \code{TRUE}, the default.}

\item{Plot}{Should the rasterized map be plotted for inspection?}
}
\value{
A rasterized, transitioned, and geo-corrected map.
}
\description{
Rasterize and transition a shapefile for input into \code{\link{Waterdist}}
}
\examples{
\dontrun{
Map <- Maptransitioner(spacetools::Delta)
}

}
\seealso{
\code{\link{Waterdist}}
}
