% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pointmover.R
\name{pointmover}
\alias{pointmover}
\title{Move points within a shapefile}
\usage{
pointmover(Data, Attribute, Shapefile)
}
\arguments{
\item{Data}{Dataframe of points with sf geometry column and an attribute column with NAs for points that need to be moved.}

\item{Attribute}{Name of the column in Data that indicates points that need to be moved (by the presence of NAs)}

\item{Shapefile}{Object of class sf representing the shapefile you want all points to fall within.}
}
\description{
Move points to closest location within shapefile
}
