% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetools.R
\docType{package}
\name{spacetools}
\alias{spacetools}
\title{spacetools: Easy spatial tools for R}
\description{
This package contains functions for easy spatial operations in R, focused on aquatic distance related applications. Internal datasets for the Sacramento San Joaquin Delta are also included to make operations very easy for this location.
}
\section{Functions}{

\itemize{
  \item \code{\link{Waterdist}}
  \item \code{\link{Pointmover}}
  \item \code{\link{Maptransitioner}}
  \item \code{\link{Pointcluster}}
}
}

\section{Internal datasets}{

\itemize{
  \item \code{\link{Delta}}
}
}

