% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Zooplankton sampling stations}
\format{a tibble with 362 rows and 3 columns
\describe{
  \item{Station}{Sampling station name}
  \item{Latitude}{Latitude in decimal degrees}
  \item{Longitude}{Longitude in decimal degrees}
}}
\usage{
stations
}
\description{
Zooplankton sampling stations in the Sacramento San Joaquin Delta from the zooper package.
}
\seealso{
\code{\link[zooper]{zooper}}
}
\keyword{datasets}
