% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waterdist.R
\name{Waterdist}
\alias{Waterdist}
\title{In-water distances between points}
\usage{
Waterdist(
  Water_map,
  Points,
  Attribute,
  Latitude_column,
  Longitude_column,
  PointID_column,
  Points_crs = 4326
)
}
\arguments{
\item{Water_map}{Object of class sf representing a map of all waterways in your region of interest}

\item{Points}{A dataframe of points with latitude and longitude}

\item{Attribute}{The name of a shapefile attribute that has values for all water bodies.}

\item{Latitude_column}{The unquoted name of the column in the Points dataframe representing Latitude.}

\item{Longitude_column}{The unquoted name of the column in the Points dataframe representing Longitude.}

\item{PointID_column}{The unquoted name of the column in the Points dataframe with the unique identifier of each point.}

\item{Points_crs}{Coordinate reference system for your points dataframe. Integer with the EPSG code or character with proj4string}
}
\value{
Distance matrix.
}
\description{
Calculate a distance matrix for a set of points based on in-water distances, using a raster-based approach
}
\keyword{distance}
\keyword{raster}
\keyword{spatial}
\keyword{water}
