% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pointcluster.R
\name{Pointcluster}
\alias{Pointcluster}
\title{Cluster nearby points}
\usage{
Pointcluster(
  Points,
  Distance,
  Latitude_column,
  Longitude_column,
  PointID_column,
  Points_crs = 4326,
  Calculation_crs = "+proj=utm +zone=10 ellps=WGS84",
  Expand = TRUE
)
}
\arguments{
\item{Points}{A dataframe of points with latitude and longitude}

\item{Distance}{Clustering distance in meters.}

\item{Latitude_column}{The unquoted name of the column in the \code{Points} dataframe representing Latitude.}

\item{Longitude_column}{The unquoted name of the column in the \code{Points} dataframe representing Longitude.}

\item{PointID_column}{The unquoted name of the column in the \code{Points} dataframe with the unique identifier of each point.}

\item{Points_crs}{Coordinate reference system for your \code{Points} dataframe. Integer with the EPSG code or character with proj4string.}

\item{Calculation_crs}{Coordinate reference system used for the calculation. If a latitude/longitude system are used, errors may be returned since some calculations assume a planar surface. Defaults to \code{Calculation_crs = "+proj=utm +zone=10 ellps=WGS84"}.}

\item{Expand}{Should data be expanded at the end? If \code{TRUE} (the default), the result will be expanded so there is 1 row per unique value of the \code{PointID_column}. If \code{FALSE} , the returned object will have 1 row per cluster and a list column with the names of each \code{PointID_column} value contained in each cluster.}
}
\value{
A tibble relating each point to its assigned cluster. The returned Latitude and Longitude values represent the mean values across all points in each cluster.
}
\description{
Cluster points within a given distance of one another
}
\examples{
library(tibble)
Points <- tibble(Latitude = c(38.07194, 38.09306, 38.11722,
                              38.11528, 38.07020, 38.09383,
                              38.11783, 38.06481, 38.11400,
                              38.06750, 38.11556),
                 Longitude = c(-122.0961, -122.0692, -122.0472,
                               -122.0519, -122.0941, -122.0697,
                               -122.0418, -122.0978, -122.0462,
                               -122.0956, -122.0424),
                 ID = c("EMP NZ022", "EMP NZ024", "EMP NZ028",
                        "EMP NZ030", "FMWT 416", "FMWT 418",
                        "FMWT 602", "TNS 418", "TNS 602",
                        "twentymm 418", "twentymm 602"))
Points_clust<-Pointcluster(Points, 1000, Latitude, Longitude, ID, Expand=TRUE)
}
