% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waterdist.R
\name{Waterdist}
\alias{Waterdist}
\title{In-water distances between points}
\usage{
Waterdist(
  Water_map,
  Points,
  Attribute,
  Latitude_column,
  Longitude_column,
  PointID_column,
  Points_crs = 4326,
  Calculation_crs = "+proj=utm +zone=10 ellps=WGS84",
  Grid_size = 75
)
}
\arguments{
\item{Water_map}{Object of class sf representing a map of all waterways in your region of interest}

\item{Points}{A dataframe of points with latitude and longitude}

\item{Attribute}{The name of a shapefile attribute that has values for all water bodies.}

\item{Latitude_column}{The unquoted name of the column in the Points dataframe representing Latitude.}

\item{Longitude_column}{The unquoted name of the column in the Points dataframe representing Longitude.}

\item{PointID_column}{The unquoted name of the column in the Points dataframe with the unique identifier of each point.}

\item{Points_crs}{Coordinate reference system for your points dataframe. Integer with the EPSG code or character with proj4string}

\item{Calculation_crs}{Coordinate reference system used for the distance calculation. If a latitude/longitude system are used, errors will be returned since these calculations assume a planar surface. Defaults to \code{Calculation_crs = "+proj=utm +zone=10 ellps=WGS84"}.}

\item{Grid_size}{Gride size (in meters) used to rasterize the map. Defaults to 75.}
}
\value{
Distance matrix.
}
\description{
Calculate a distance matrix for a set of points based on in-water distances, using a raster-based approach
}
\examples{
Points <- data.frame(Latitude = c(38.04813, 38.05920, 37.94900, 38.23615, 38.47387),
Longitude = c(-121.9149, -121.8684, -121.5591, -121.6735, -121.5844),
ID = c("FMWT 508", "FMWT 513", "FMWT 915", "FMWT 723", "FMWT 796"))
distance<-Waterdist(spacetools::Delta, Points, Shape_Area, Latitude, Longitude, ID)
}
\keyword{distance}
\keyword{raster}
\keyword{spatial}
\keyword{water}
